Add online processor based on queue
